refactor: move toast to ui components
