+++ project/components/modals/create-project-modal.tsx
@@ -4,7 +4,7 @@ import { useState } from "react";
 import Button from "../button";
 import Input from "../input";
 import Textarea from "../textarea";
-import { showErrorToast, showSuccessToast } from "../toast";
+import { showErrorToast, showSuccessToast } from "../ui/toast";
 import Modal from "./modal";
 
 interface CreateProjectModalProps {
