+++ project/components/modals/delete-project-modal.tsx
@@ -4,7 +4,7 @@ import { redirect, RedirectType } from "next/navigation";
 import { ChangeEvent, useState } from "react";
 import Button from "../button";
 import Input from "../input";
-import { showErrorToast, showSuccessToast } from "../toast";
+import { showErrorToast, showSuccessToast } from "../ui/toast";
 import Modal from "./modal";
 
 interface DeleteProjectModalProps {
