+++ project/components/modals/edit-project-modal.tsx
@@ -38,7 +38,7 @@ import { useEffect, useState } from "react";
 import Button from "../button";
 import Input from "../input";
 import Textarea from "../textarea";
-import { showErrorToast, showSuccessToast } from "../toast";
+import { showErrorToast, showSuccessToast } from "../ui/toast";
 import Modal from "./modal";
 
 interface EditProjectModalProps {
