+++ project/components/modals/update-list-modal.tsx
@@ -11,7 +11,7 @@ import {
 import Button from "../button";
 import Input from "../input";
 import Radio from "../radio";
-import { showErrorToast, showSuccessToast } from "../toast";
+import { showErrorToast, showSuccessToast } from "../ui/toast";
 import Modal from "./modal";
 
 interface UpdateListModalProps {
