+++ project/components/select-project-members.tsx
@@ -4,7 +4,7 @@ import { getProjectMembersOptions } from "@/lib/actions/teams";
 import { ProjectSchema } from "@/types/projects";
 import { UserSchema } from "@/types/users";
 import { ChangeEvent, useCallback, useEffect, useState } from "react";
-import { showErrorToast } from "./toast";
+import { showErrorToast } from "./ui/toast";
 import UserCheckbox from "./user-checkbox";
 
 interface SelectMembersProps {
